backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyMuPDF
anthropic
redis
python-dotenv
pydantic-settings
pydantic
